refactor: more unity core

+++ CppHaxSDK/cheat/globals.cpp
@@ -6,6 +6,6 @@
 namespace globals {
 	bool	g_visible = true;
 	VK_KEY	g_showMenuKey = VK_OEM_3;
-	long	g_screenWidth = 0;
-	long	g_screenHeight = 0;
+	float	g_screenWidth = 0;
+    float	g_screenHeight = 0;
 }
