+++ CppHaxSDK/cheat/globals.h
@@ -5,6 +5,6 @@ typedef int VK_KEY;
 namespace globals {
 	extern bool		g_visible;
 	extern VK_KEY	g_showMenuKey;
-	extern long		g_screenWidth;
-	extern long		g_screenHeight;
+	extern float	g_screenWidth;
+	extern float	g_screenHeight;
 }
